Add files via upload
This Java Program illustrates how to read data from a text File using Scanner Class (Note: 'data.txt' is used in conjunction with DataFileReader.java. The contents of the data file will be outputted to the console for the user to view. To change the data file, simply add it to  the program directory and rename the file pointer or list out the path directory.