Add files via upload
This Java Program demonstrates how object-orientated programming in conjunction with class constructors can be used to create an ATM for a Bank with full scale transactions such as withdrawal, deposits, and interest rates factored in. Each Bank Savings and Checking Accounts comes with unique user IDs and passwords that can be updated if needed. A bonus feature is to let users choose what kind of dollar bills they want when making a withdrawal.